requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
googlemaps==4.10.0
geopy==2.4.0
folium==0.15.1
beautifulsoup4==4.12.2
lxml==4.9.3 